now is an f64

+++ src/aircraft.rs
@@ -3,6 +3,6 @@ use serde_json::Value;
 
 #[derive(Deserialize, Debug)]
 pub struct AircraftFile {
-    pub now: f32,
+    pub now: f64,
     pub aircraft: Vec<Value>
 }
